remove tooltip popper dependency from server-side imports
