+++ components/common/CharmEditor/components/inlineComment/index.ts
@@ -1,4 +1,3 @@
-export * from './inlineComment.specs';
 export * from './inlineComment.plugins';
 export * from './inlineComment.constants';
 export * from './inlineComment.utils';
