+++ components/common/CharmEditor/components/inlineVote/index.ts
@@ -1,4 +1,3 @@
-export * from './inlineVote.specs';
 export * from './inlineVote.plugins';
 export * from './inlineVote.constants';
 export * from './inlineVote.utils';
