+++ components/common/CharmEditor/components/mention/index.ts
@@ -1,5 +1,4 @@
 export * from './mention.constants';
 export * from './mention.plugins';
 export * from './mention.interfaces';
-export * from './mention.specs';
 export * from './mention.utils';
