+++ components/common/CharmEditor/specRegistry.ts
@@ -19,10 +19,10 @@ import { hardBreakSpec } from './components/hardBreak';
 import * as heading from './components/heading';
 import * as horizontalRule from './components/horizontalRule';
 import * as iframe from './components/iframe';
-import * as inlineComment from './components/inlineComment';
+import * as inlineComment from './components/inlineComment/inlineComment.specs';
 import * as inlineDatabase from './components/inlineDatabase';
 import { spec as inlinePaletteSpecs } from './components/inlinePalette/inlinePaletteSpec';
-import * as inlineVote from './components/inlineVote';
+import * as inlineVote from './components/inlineVote/inlineVote.specs';
 import { spec as linkSpec } from './components/link/link.specs';
 import { linkedPageSpec } from './components/linkedPage/linkedPage.specs';
 import * as listItem from './components/listItem/listItem';
@@ -30,7 +30,7 @@ import { spec as bulletListItemSpec } from './components/listItemNew/bulletListS
 import { spec as listItemSpec } from './components/listItemNew/listItemSpecs';
 import { spec as markSpec } from './components/listItemNew/markSpec';
 import { spec as orderedListItemSpec } from './components/listItemNew/orderedListSpecs';
-import { mentionSpecs } from './components/mention';
+import { mentionSpecs } from './components/mention/mention.specs';
 import { nestedPageSpec } from './components/nestedPage/nestedPage.specs';
 import * as nft from './components/nft/nft.specs';
 import * as orderedList from './components/orderedList';
