+++ lib/prosemirror/insertLinkedPage.ts
@@ -2,7 +2,7 @@ import type { EditorView, PluginKey } from '@bangle.dev/pm';
 import { TextSelection } from '@bangle.dev/pm';
 import { rafCommandExec } from '@bangle.dev/utils';
 
-import { hideSuggestionsTooltip } from 'components/common/CharmEditor/components/@bangle.dev/tooltip/suggestTooltip';
+import { hideSuggestionsTooltip } from 'components/common/CharmEditor/components/@bangle.dev/tooltip/suggestTooltipSpec';
 import { replaceSuggestionMarkWith } from 'components/common/CharmEditor/components/inlinePalette/inlinePalette';
 import { linkedPageNodeName } from 'components/common/CharmEditor/components/linkedPage/linkedPage.constants';
 import { type LinkedPagePluginState } from 'components/common/CharmEditor/components/linkedPage/linkedPage.interfaces';
