+++ tsconfig.json
@@ -58,5 +58,4 @@
     }
   },
   "exclude": ["node_modules"],
-  "include": ["**/*.ts", "**/*.tsx"]
 }
